Distinguish between prompt view and prompt interaction events
