+++ amplify/src/test/java/com/github/stkent/amplify/tracking/PromptViewEventTest.java
@@ -37,104 +37,6 @@ public class PromptViewEventTest extends BaseTest {
     }
 
     @Test
-    public void testThatUserIndicatedPositiveOpinionEventHasCorrectTrackingKey() {
-        // Arrange
-        final IEvent userIndicatedPositiveOpinionEvent
-                = PromptViewEvent.USER_INDICATED_POSITIVE_OPINION;
-
-        // Assert
-        Assert.assertEquals(
-                "User indicated positive opinion event should have correct tracking key",
-                userIndicatedPositiveOpinionEvent.getTrackingKey(),
-                "USER_INDICATED_POSITIVE_OPINION");
-    }
-
-    @Test
-    public void testThatUserIndicatedCriticalOpinionEventHasCorrectTrackingKey() {
-        // Arrange
-        final IEvent userIndicatedCriticalOpinionEvent
-                = PromptViewEvent.USER_INDICATED_CRITICAL_OPINION;
-
-        // Assert
-        Assert.assertEquals(
-                "User indicated critical opinion event should have correct tracking key",
-                userIndicatedCriticalOpinionEvent.getTrackingKey(),
-                "USER_INDICATED_CRITICAL_OPINION");
-    }
-
-    @Test
-    public void testThatUserGavePositiveFeedbackEventHasCorrectTrackingKey() {
-        // Arrange
-        final IEvent appCrashedEvent = PromptViewEvent.USER_GAVE_POSITIVE_FEEDBACK;
-
-        // Assert
-        Assert.assertEquals(
-                "User gave positive feedback event should have correct tracking key",
-                appCrashedEvent.getTrackingKey(),
-                "USER_GAVE_POSITIVE_FEEDBACK");
-    }
-
-    @Test
-    public void testThatUserGaveCriticalFeedbackEventHasCorrectTrackingKey() {
-        // Arrange
-        final IEvent appCrashedEvent = PromptViewEvent.USER_GAVE_CRITICAL_FEEDBACK;
-
-        // Assert
-        Assert.assertEquals(
-                "User gave critical feedback event should have correct tracking key",
-                appCrashedEvent.getTrackingKey(),
-                "USER_GAVE_CRITICAL_FEEDBACK");
-    }
-
-    @Test
-    public void testThatUserDeclinedPositiveFeedbackEventHasCorrectTrackingKey() {
-        // Arrange
-        final IEvent appCrashedEvent = PromptViewEvent.USER_DECLINED_POSITIVE_FEEDBACK;
-
-        // Assert
-        Assert.assertEquals(
-                "User declined positive feedback event should have correct tracking key",
-                appCrashedEvent.getTrackingKey(),
-                "USER_DECLINED_POSITIVE_FEEDBACK");
-    }
-
-    @Test
-    public void testThatUserDeclinedCriticalFeedbackEventHasCorrectTrackingKey() {
-        // Arrange
-        final IEvent appCrashedEvent = PromptViewEvent.USER_DECLINED_CRITICAL_FEEDBACK;
-
-        // Assert
-        Assert.assertEquals(
-                "User declined critical feedback event should have correct tracking key",
-                appCrashedEvent.getTrackingKey(),
-                "USER_DECLINED_CRITICAL_FEEDBACK");
-    }
-
-    @Test
-    public void testThatUserGaveFeedbackEventHasCorrectTrackingKey() {
-        // Arrange
-        final IEvent userGaveFeedbackEvent = PromptViewEvent.USER_GAVE_FEEDBACK;
-
-        // Assert
-        Assert.assertEquals(
-                "User gave feedback event should have correct tracking key",
-                userGaveFeedbackEvent.getTrackingKey(),
-                "USER_GAVE_FEEDBACK");
-    }
-
-    @Test
-    public void testThatUserDeclinedFeedbackEventHasCorrectTrackingKey() {
-        // Arrange
-        final IEvent userDeclinedFeedbackEvent = PromptViewEvent.USER_DECLINED_FEEDBACK;
-
-        // Assert
-        Assert.assertEquals(
-                "User declined feedback event should have correct tracking key",
-                userDeclinedFeedbackEvent.getTrackingKey(),
-                "USER_DECLINED_FEEDBACK");
-    }
-
-    @Test
     public void testThatThanksShownEventHasCorrectTrackingKey() {
         // Arrange
         final IEvent thanksShownEvent = PromptViewEvent.THANKS_SHOWN;
